Refactor to SceneOptimizer
